chore: refactor product category and VAT group seeders
