+++ test/unit/service/container-service.ts
@@ -30,7 +30,6 @@ import Swagger from '../../../src/start/swagger';
 import ContainerService from '../../../src/service/container-service';
 import {
   seedContainers, seedProducts, seedPointsOfSale,
-  seedProductCategories, seedVatGroups,
 } from '../../seed-legacy';
 import Container from '../../../src/entity/container/container';
 import { ContainerResponse, ContainerWithProductsResponse } from '../../../src/controller/response/container-response';
@@ -49,7 +48,7 @@ import { finishTestDB } from '../../helpers/test-helpers';
 import PointOfSale from '../../../src/entity/point-of-sale/point-of-sale';
 import sinon from 'sinon';
 import { PointOfSaleWithContainersResponse } from '../../../src/controller/response/point-of-sale-response';
-import { UserSeeder } from '../../seed';
+import { ProductCategorySeeder, UserSeeder, VatGroupSeeder } from '../../seed';
 
 /**
  * Test if all the container responses are part of the container set array.
@@ -114,8 +113,8 @@ describe('ContainerService', async (): Promise<void> => {
     await truncateAllTables(connection);
 
     const users = await new UserSeeder().seedUsers();
-    const categories = await seedProductCategories();
-    const vatGroups = await seedVatGroups();
+    const categories = await new ProductCategorySeeder().seedProductCategories();
+    const vatGroups = await new VatGroupSeeder().seedVatGroups();
     const { productRevisions } = await seedProducts(users, categories, vatGroups);
     const { containers, containerRevisions } = await seedContainers(users, productRevisions);
     const { pointsOfSale, pointOfSaleRevisions } = await seedPointsOfSale(users, containerRevisions);
