+++ test/unit/service/seller-payout-service.ts
@@ -23,11 +23,9 @@ import database from '../../../src/database/database';
 import {
   seedContainers,
   seedPointsOfSale,
-  seedProductCategories,
   seedProducts,
   seedTransactions,
   seedTransfers,
-  seedVatGroups,
 } from '../../seed-legacy';
 import { finishTestDB } from '../../helpers/test-helpers';
 import SubTransaction from '../../../src/entity/transactions/sub-transaction';
@@ -37,7 +35,7 @@ import SellerPayoutService, { CreateSellerPayoutParams } from '../../../src/serv
 import { calculateBalance } from '../../helpers/balance';
 import { DineroObjectRequest } from '../../../src/controller/request/dinero-request';
 import dinero from 'dinero.js';
-import { SellerPayoutSeeder, UserSeeder } from '../../seed';
+import { ProductCategorySeeder, SellerPayoutSeeder, UserSeeder, VatGroupSeeder } from '../../seed';
 
 describe('SellerPayoutService', () => {
   let ctx: {
@@ -53,8 +51,8 @@ describe('SellerPayoutService', () => {
     const connection = await database.initialize();
     const users = await new UserSeeder().seedUsers();
 
-    const categories = await seedProductCategories();
-    const vatGroups = await seedVatGroups();
+    const categories = await new ProductCategorySeeder().seedProductCategories();
+    const vatGroups = await new VatGroupSeeder().seedVatGroups();
     const { productRevisions } = await seedProducts(users, categories, vatGroups);
     const { containerRevisions } = await seedContainers(users, productRevisions);
     const { pointOfSaleRevisions } = await seedPointsOfSale(users, containerRevisions);
