+++ test/unit/service/transfer-service.ts
@@ -31,17 +31,15 @@ import Swagger from '../../../src/start/swagger';
 import {
   seedContainers, seedFines,
   seedInvoices, seedPayoutRequests, seedPointsOfSale,
-  seedProductCategories,
   seedProducts, seedStripeDeposits,
   seedTransactions,
   seedTransfers,
-  seedVatGroups,
 } from '../../seed-legacy';
 import DineroTransformer from '../../../src/entity/transformer/dinero-transformer';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
 import VatGroup from '../../../src/entity/vat-group';
-import { UserSeeder } from '../../seed';
+import { ProductCategorySeeder, UserSeeder, VatGroupSeeder } from '../../seed';
 
 describe('TransferService', async (): Promise<void> => {
   let ctx: {
@@ -60,8 +58,8 @@ describe('TransferService', async (): Promise<void> => {
     const end = new Date('2001-02-12T01:57:45.271Z');
 
     const users = await new UserSeeder().seedUsers();
-    const vatGroups = await seedVatGroups();
-    const categories = await seedProductCategories();
+    const vatGroups = await new VatGroupSeeder().seedVatGroups();
+    const categories = await new ProductCategorySeeder().seedProductCategories();
     const { productRevisions } = await seedProducts(users, categories, vatGroups);
     const { containerRevisions } = await seedContainers(users, productRevisions);
     const { pointOfSaleRevisions } = await seedPointsOfSale(users, containerRevisions);
