+++ test/unit/service/vat-group-service.ts
@@ -26,16 +26,14 @@ import Database from '../../../src/database/database';
 import {
   seedContainers,
   seedPointsOfSale,
-  seedProductCategories,
   seedProducts,
   seedTransactions,
-  seedVatGroups,
 } from '../../seed-legacy';
 import VatGroupService from '../../../src/service/vat-group-service';
 import { VatDeclarationResponse } from '../../../src/controller/response/vat-group-response';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
-import { UserSeeder } from '../../seed';
+import { ProductCategorySeeder, UserSeeder, VatGroupSeeder } from '../../seed';
 
 describe('VatGroupService', () => {
   let ctx: {
@@ -50,8 +48,8 @@ describe('VatGroupService', () => {
     const connection = await Database.initialize();
     await truncateAllTables(connection);
     const users = await new UserSeeder().seedUsers();
-    const vatGroups = await seedVatGroups();
-    const categories = await seedProductCategories();
+    const vatGroups = await new VatGroupSeeder().seedVatGroups();
+    const categories = await new ProductCategorySeeder().seedProductCategories();
     const { productRevisions } = await seedProducts(users, categories, vatGroups, 100);
     const { containerRevisions } = await seedContainers(users, productRevisions);
     const { pointOfSaleRevisions } = await seedPointsOfSale(users, containerRevisions);
