+++ test/unit/subscribe/transaction-subscriber.ts
@@ -25,9 +25,7 @@ import Database from '../../../src/database/database';
 import {
   seedContainers,
   seedPointsOfSale,
-  seedProductCategories,
   seedProducts, seedTransactions, seedTransfers,
-  seedVatGroups,
 } from '../../seed-legacy';
 import { calculateBalance } from '../../helpers/balance';
 import ProductRevision from '../../../src/entity/product/product-revision';
@@ -41,7 +39,7 @@ import TransactionService from '../../../src/service/transaction-service';
 import BalanceService from '../../../src/service/balance-service';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
-import { UserSeeder } from '../../seed';
+import { ProductCategorySeeder, UserSeeder, VatGroupSeeder } from '../../seed';
 
 describe('TransactionSubscriber', () => {
   let ctx: {
@@ -77,8 +75,8 @@ describe('TransactionSubscriber', () => {
     } as User;
 
     const users = await new UserSeeder().seedUsers();
-    const categories = await seedProductCategories();
-    const vatGroups = await seedVatGroups();
+    const categories = await new ProductCategorySeeder().seedProductCategories();
+    const vatGroups = await new VatGroupSeeder().seedVatGroups();
     const { productRevisions } = await seedProducts([adminUser], categories, vatGroups);
     const { containerRevisions } = await seedContainers([adminUser], productRevisions);
     const { pointOfSaleRevisions } = await seedPointsOfSale([adminUser], containerRevisions);
